backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic[email]==2.6.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.2
elasticsearch==8.12.1
redis==5.0.3
pika==1.3.2
scikit-learn==1.4.1.post1
tensorflow==2.15.0
nltk==3.8.1
spacy==3.7.4
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
pandas==2.2.1
gensim==4.3.2
requests==2.31.0
aiohttp==3.9.3
httpx==0.26.0
jinja2==3.1.3
itsdangerous==2.1.2 